+++ cmd/zarb/main.go
@@ -21,11 +21,11 @@ func zarb() *cli.Cli {
 		k.Command("verify", "Verify a signature", key.Verify())
 		k.Command("change-auth", "Change the passphrase of a keyfile", key.ChangeAuth())
 	})
-	app.Command("tx", "Create raw transaction", func(k *cli.Cmd) {
-		k.Command("bond", "Generate raw bond transaction", tx.BondTx())
-		k.Command("send", "Generate raw send transaction", tx.SendTx())
-		k.Command("unbond", "Generate raw unbond transaction", tx.UnbondTx())
-		k.Command("withdraw", "Generate raw withdraw transaction", tx.WithdrawTx())
+	app.Command("tx", "Create, sign and publish a transaction", func(k *cli.Cmd) {
+		k.Command("bond", "Create, sign and publish a bond transaction", tx.BondTx())
+		k.Command("send", "Create, sign and publish a send transactio", tx.SendTx())
+		k.Command("unbond", "Create, sign and publish an unbond transaction", tx.UnbondTx())
+		k.Command("withdraw", "Create, sign and publish a withdraw transaction", tx.WithdrawTx())
 	})
 	app.Command("version", "Print the zarb version", Version())
 	return app