+++ execution/executor/bond.go
@@ -42,6 +42,7 @@ func (e *BondExecutor) Execute(trx *tx.Tx, sb sandbox.Sandbox) error {
 	bonderAcc.IncSequence()
 	bonderAcc.SubtractFromBalance(pld.Stake + trx.Fee())
 	val.AddToStake(pld.Stake)
+	val.UpdateLastBondingHeight(sb.CurrentHeight())
 
 	sb.UpdateAccount(bonderAcc)
 	sb.UpdateValidator(val)