+++ execution/executor/sortition.go
@@ -25,7 +25,7 @@ func (e *SortitionExecutor) Execute(trx *tx.Tx, sb sandbox.Sandbox) error {
 	if val.Power() == 0 {
 		return errors.Errorf(errors.ErrInvalidTx, "Validator has no Power to be in committee")
 	}
-	if sb.CurrentHeight()-val.BondingHeight() < 2*sb.CommitteeSize() {
+	if sb.CurrentHeight()-val.LastBondingHeight() < 2*sb.CommitteeSize() {
 		return errors.Errorf(errors.ErrInvalidTx, "in bonding period")
 	}
 	if !sb.VerifySortition(trx.Stamp(), pld.Proof, val) {