+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 1.1.0
+- Updating last bonding height for validators
+- Refactor Transactions in CMD
+- ADd support for pprof
+
 ## 1.0.0
 - Refactoring Consensus module
 - Refactoring Sync module